requirements.txt
Django==4.2.13
djangorestframework==3.14.0
drf_nested_routers==0.94.1
matplotlib==3.9.1
pandas==2.2.2
plotly==5.22.0
reportlab==4.2.2
twilio==9.2.3
